IFAKE_WebApp/requirements.txt
Django==1.11.29
Django==3.2.1
keras==2.6.0
matplotlib==3.3.3
matplotlib==3.3.4
numpy==1.19.5
opencv_contrib_python==4.5.5.62
pandas==1.1.4
pandas==1.2.3
pdf2image==1.16.0
Pillow==8.1.2
Pillow==9.5.0
scikit_image==0.18.1
scikit_learn==0.24.1
scipy==1.6.1
streamlit==0.81.1
tensorflow==2.6.0
tqdm==4.59.0
~pencv_python==4.5.3.56
